+++ x-pack/packages/kbn-elastic-assistant-common/impl/language_models/helpers.test.ts
@@ -7,9 +7,6 @@
 
 import { Message } from '../schemas';
 import { getMessageContentAndRole } from './helpers';
-import { getCombinedMessage } from '@kbn/elastic-assistant/impl/assistant/prompt/helpers';
-import type { SelectedPromptContext } from '@kbn/elastic-assistant/impl/assistant/prompt_context/types';
-import type { Prompt } from '@kbn/elastic-assistant';
 
 describe('helpers', () => {
   describe('getMessageContentAndRole', () => {
@@ -25,95 +22,6 @@ describe('helpers', () => {
 
         expect(result).toEqual(expectedOutput);
       });
-    });
-  });
-  describe('getCombinedMessage', () => {
-    const defaultGetAnonymizedValue = jest.fn((args) => args.rawValue);
-
-    const mockPrompt: Prompt = {
-      content: 'This is a system prompt',
-      id: '123',
-      name: 'prompt name',
-      promptType: 'system',
-    };
-
-    const mockPromptContexts: Record<string, SelectedPromptContext> = {
-      context1: {
-        promptContextId: 'context1',
-        rawData: 'This is raw data for context 1',
-        replacements: {},
-      },
-      context2: {
-        promptContextId: 'context2',
-        rawData: 'This is raw data for context 2',
-        replacements: {},
-      },
-    };
-
-    const mockCurrentReplacements = {
-      currentReplacement1: 'value1',
-      currentReplacement2: 'value2',
-    };
-    const defaultProps = {
-      currentReplacements: mockCurrentReplacements,
-      getAnonymizedValue: defaultGetAnonymizedValue,
-      isNewChat: true,
-      promptText: 'This is a user prompt',
-      selectedPromptContexts: mockPromptContexts,
-      selectedSystemPrompt: mockPrompt,
-    };
-
-    beforeEach(() => {
-      jest.clearAllMocks();
-    });
-
-    it('should return the correct combined message for a new chat', () => {
-      const result = getCombinedMessage(defaultProps);
-
-      expect(result.content).toEqual(
-        `This is a system prompt\n\nCONTEXT:\n\"\"\"\nThis is raw data for context 1\n\"\"\"\n,CONTEXT:\n\"\"\"\nThis is raw data for context 2\n\"\"\"\n\nThis is a user prompt`
-      );
-    });
-
-    it('should return the correct combined message for an existing chat', () => {
-      const result = getCombinedMessage({
-        ...defaultProps,
-        isNewChat: false,
-      });
-
-      expect(result.content).toEqual(
-        `CONTEXT:\n\"\"\"\nThis is raw data for context 1\n\"\"\"\n,CONTEXT:\n\"\"\"\nThis is raw data for context 2\n\"\"\"\n\nThis is a user prompt`
-      );
-    });
-
-    it('should return the correct combined message for a new chat without prompt context', () => {
-      const result = getCombinedMessage({
-        ...defaultProps,
-        selectedPromptContexts: {},
-      });
-
-      expect(result.content).toEqual(`This is a system prompt\n\nThis is a user prompt`);
-    });
-
-    it('should return the correct combined message for a new chat without system context', () => {
-      const result = getCombinedMessage({
-        ...defaultProps,
-        selectedSystemPrompt: { ...mockPrompt, content: '' },
-      });
-
-      expect(result.content).toEqual(
-        `CONTEXT:\n\"\"\"\nThis is raw data for context 1\n\"\"\"\n,CONTEXT:\n\"\"\"\nThis is raw data for context 2\n\"\"\"\n\nThis is a user prompt`
-      );
-    });
-
-    it('should remove extra spaces when there is no prompt content or system prompt', () => {
-      const result = getCombinedMessage({
-        ...defaultProps,
-        selectedPromptContexts: {},
-        selectedSystemPrompt: { ...mockPrompt, content: '' },
-      });
-
-      expect(result.content).toEqual(`This is a user prompt`);
     });
   });
 });
